+++ src/main/java/insane96mcp/insanesurvivaloverhaul/mixin/module/hungerhealth/exhaustion/PlayerMixin_Exhaustion.java
@@ -1,4 +1,4 @@
-package insane96mcp.insanesurvivaloverhaul.mixin.module.hungerhealth;
+package insane96mcp.insanesurvivaloverhaul.mixin.module.hungerhealth.exhaustion;
 
 import com.llamalad7.mixinextras.injector.ModifyExpressionValue;
 import insane96mcp.insanelib.core.feature.Feature;