+++ changelog.md
@@ -1,5 +1,7 @@
 ## Upcoming
 * Sprinting consumes much more hunger (5x)
+  * Also heavily increased sprint jumping exhaustion
+  * Reduced non-sprint jumping exhaustion
 
 ## 5.4.3.0
 * Added Anvil Crafting